requirements.txt
langchain==0.0.344
pypdf==4.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2

# For Streamlit web interface
streamlit==1.28.0

